http://2timexmedia.com
Here are cloves health uses as follows:
Acne buster. 

Because of the eugenol found in cloves, cloves can be used to help prevent acne breakouts. You can make a mask with ground cloves, honey and a few drops of lemon juice. Keep it on your face for around 20 minutes and then rinse.

Mouthwash. 

If you’re tired of the chemical aftertaste that mouthwashes leave in your mouth, you can switch to cloves as an all-natural alternative. Cloves not only will freshen your breath, but will give you anti-inflammatory and antibacterial benefits as well. Natural clove mouthwashes usually consist of a mixture of equal amounts of water, cloves and other herbs like rosemary and mint.

Toothache remedy. 

Clove oil can be used as a natural painkiller for toothaches. This is because of its natural anesthetic property that helps alleviate pain and discomfort that arise from cavities and other dental and gum problems.

Altitude sickness relief. 

Mountaineers, climbers and other sportsmen usually suffer from altitude sickness because of the sudden change in atmospheric pressure. Taking 2 cups of a clove infusion can help relieve altitude sickness by thinning the blood and improving the oxygen supply to the brain.