Update README.md
Improves Digestion
Using cloves in your meals will add a burst of flavor to them and also aid the digestive process by stimulating the secretion of digestive enzymes. They eliminate gas that collects in the digestive tracts which makes them a good cure for flatulence. They can be used in powder form with honey as a home-treatment for stomach disorders. As a result, you’re less likely to feel nauseous because of indigestion.

Antibacterial Properties:
Cloves have been tested for their antibacterial properties against a number of human pathogens. The extracts of cloves were potent enough to kill those pathogens. Clove extracts are also effective against the specific bacteria that spread cholera.

Anti-Carcinogenic
Studies show that cloves can act as an effective anti-carcinogenic which is a substance that can protect that body against the growth of cancers by inhibiting them. This was applied specifically to lung cancer, and results showed that cloves stopped the growth completely.